HumShapes fixed, HUM2 now starts from Shape 14 then HUM3 starts from 28.  This should now sort the issue with the last armour not loading..
New Monster FlamingMutant added,Race 221, RaceImg 136 Appr 237.

This guy is very uniquie, the AI can be heavily configuired by the Server owner in a Config file called !MonsterConfig.txt located in the mir200 folder.  Have a play with the monster in game first then look at the config file and adjust the settings to suit your needs.  His meteor attack damage is based on MC so adjust that in the Database.  His critical attack is his DC, but multiplied by the setting in !MonsterConfig.txt.

Er i think that's it for this guy, have a play with him and drop some feedback on the forum.  Thanks.