Skills added Hero AI Updated to use new skill New item Stuff used in StdItems, in Description put Player for player only use or Hero for hero only use. Pots now can use Level requirements.
MagicDB can now let normal skills be used for Hero
in the description menu put 'Both', and this will allow a hero to learn that spell.

Current server owners will have to add a script in to Re add there users Skills since the ID's have been player around with.  Sorry for any trouble this causes you, as this couldnt be helped.

Download DB on Google SVN